Separate json headers from pybind headers to remove standalone dependency on pybind
