+++ qiskit/providers/aer/backends/wrappers/statevector_bindings.cc
@@ -2,8 +2,7 @@
 
 #include "simulators/statevector/statevector_controller.hpp"
 #include "simulators/controller_execute.hpp"
-
-#include "framework/json.hpp"
+#include "framework/pybind_json.hpp"
 
 PYBIND11_MODULE(statevector_controller_wrapper, m) {
     m.def("statevector_controller_execute_json", &AER::controller_execute_json<AER::Simulator::StatevectorController>, "instance of controller_execute for StatevectorController");
