+++ qiskit/providers/aer/backends/wrappers/unitary_bindings.cc
@@ -2,8 +2,7 @@
 
 #include "simulators/unitary/unitary_controller.hpp"
 #include "simulators/controller_execute.hpp"
-
-#include "framework/json.hpp"
+#include "framework/pybind_json.hpp"
 
 PYBIND11_MODULE(unitary_controller_wrapper, m) {
     m.def("unitary_controller_execute_json", &AER::controller_execute_json<AER::Simulator::UnitaryController>, "instance of controller_execute for UnitaryController");
